moved EditList into ItemRow component. Edited so new items can appear without refresh. Debugged enter-to-save issue in ItemRow.
