+++ app/src/ListItems.js
@@ -1,18 +1,13 @@
-import React, { Fragment, useEffect, useState } from "react";
+import React, { useState, useEffect } from "react";
 
 import AddItem from "./AddItemForm";
-import { EditItem, Checkbox } from "./EditItem";
+import ItemRow from "./ItemRow";
 import * as apiClient from "./apiClient";
 
 // This is view for one list.
 const ListItems = () => {
   const [items, setItems] = useState([]);
 
-  // const loadItems = async () => setItems(await apiClient.getItems());
-
-  // useEffect(() => {
-  //   loadItems();
-  // }, []);
   async function getItems() {
     const itemsArray = await apiClient.getItems();
 
@@ -21,6 +16,7 @@ const ListItems = () => {
   useEffect(() => {
     getItems();
   }, []);
+
   // delete item function
   async function deleteItem(id) {
     apiClient.deleteItem(id);
@@ -29,12 +25,14 @@ const ListItems = () => {
     setItems(items.filter((item) => item.id !== id));
   }
 
+  const onAdd = (item) => setItems([...items, item]);
+
   return (
     <>
       <div className="body">
         <h1>Main List</h1>
         <br />
-        <AddItem />
+        <AddItem onAdd={onAdd} />
         <table className="table table-hover mt-5">
           <thead>
             <tr className="header-row">
@@ -54,36 +52,7 @@ const ListItems = () => {
               </tr>
   */}
             {items.map((item) => (
-              <tr key={item.id} className="item-row">
-                <td>
-                  <Checkbox />
-                </td>
-                <EditItem item={item} />
-                <td>Recurring Option</td>
-                <td>
-                  <button
-                    className="btn btn-danger"
-                    onClick={() => deleteItem(item.id)}
-                  >
-                    Delete
-                    {/* Delete icon. Removed since emojis are not ARIA - compliant. Tested and works.*/}
-                    {/* <span
-                    role="img"
-                    id={`id${item.id}`}
-                    aria-label="Delete item button"
-                    onClick={() => {
-                      deleteItem(item.id);
-                    }}
-                    onKeyPress={() => {
-                      deleteItem(item.id);
-                    }}
-                    aria-hidden="true"
-                  >
-                    &#10060;
-                  </span> */}
-                  </button>
-                </td>
-              </tr>
+              <ItemRow item={item} deleteItem={deleteItem} />
             ))}
           </tbody>
         </table>
